Construction Level Work (New Ending)
Some tweaks to the level layout. End section redone with moving platforms.

Far From prefect, but it's getting better.

Still some jankiness probably due to the checkpoint's misaligning the player when teleporting/repsawning them (because of the Respawn Point not being in-line with the rest of the level), which can mean they might not touch cogs or hit & activate JumpPads when they should.

Also added a temporary placeholder background image (made using Stable Diffusion)