Adjustments and fixes
o Separate Camera Actor Idea has been dropped in favour of a Spring Arm on the player. It gives us more control over the rotation and distance of the camera

o Skydome has been added to the tutorial level

o Imported Player character mesh. (There are a lot of issues to fix as its currently unavailable)