Small Updates
Some patches have been made to the dialogue system so that the text changes aren't seen by the player. The input is now also disabled to allow the player time to read what they are being told. All that needs done now is the implementation of the tutorial lines.

An update has been made to the skydome to allow better UV tiling so we can shrink the texture